rag/requirements.txt
# Core web framework
fastapi>=0.127.0
uvicorn[standard]>=0.40.0
pydantic>=2.12.0
python-dotenv>=1.0.1

# LangChain and OpenAI
langchain>=1.2.0
langchain-openai>=1.1.0
langchain-core>=1.2.0
langchain-community>=0.4.0
openai>=2.14.0
tiktoken>=0.12.0

# Vector database and search
numpy>=2.4.0
faiss-cpu>=1.7.4
chromadb>=1.3.0

# Web scraping and parsing
beautifulsoup4>=4.14.0
lxml>=6.0.0
requests>=2.32.0
httpx>=0.28.0

# Production server and rate limiting
gunicorn>=21.2.0
slowapi>=0.1.9

# Testing
pytest>=9.0.0
pytest-asyncio>=0.25.0

# Development tools
ruff>=0.14.0
mypy>=1.19.0
pip-review>=1.3.0

# Optional production monitoring (can remove if not needed)
redis>=5.0.0
psutil>=5.9.0
prometheus-client>=0.19.0
structlog>=23.2.0
